add saving files

+++ ventilated_facade_items/ventilated_facade_items.controller.js
@@ -1,14 +1,25 @@
 const { VentilatedFacadesItemsDAO } = require('./ventilated_facade_items.DAO')
 const sharp = require('sharp')
 const fs = require('fs')
+const path = require('path');
+// const directoryPath = './static/facadesItems/';
+const os = require('os');        // await sharp(req.file.path)
+// .toFile(path.resolve(__dirname, `./static/facadesItems/${req.file.originalname}`))
+
 
 class VentilatedFacadesItemsItemsController {
     async createVentilatedFacadeItem(req, res) {
         const { ventilatedFacadeId } = req.body
+        console.log(os.userInfo().username);
+        if (!req.file) {
+            res.status(400).send({message: 'file was not transferred'})
+            return
+        }
         await sharp(req.file.path)
-            .toFile(`./static/facadesItems/${req.file.originalname}`)
+            .toFile(`/usr/src/app/static/facadesItems/${req.file.originalname}`)
 
-        const url = `http://localhost:8000/static/facadesItems/${req.file.originalname}`
+        const url = `https://frolfasd.ru/static/facadesItems/${req.file.originalname}`
+        console.log('изменение!!!')
         fs.unlink(req.file.path, () => { // Для удаления закодированных файлов после использования
             console.log(req.file.path)
         })
@@ -24,6 +35,39 @@ class VentilatedFacadesItemsItemsController {
                 }
             });
     }
+
+    // async createVentilatedFacadeItem(req, res) {
+    //     const { ventilatedFacadeId } = req.body
+    
+    //     try {
+    //         // Проверяем существование каталога и создаем его, если нет
+    //         await fs.access(directoryPath);
+    //     } catch (error) {
+    //         if (error.code === 'ENOENT') {
+    //             await fs.mkdir(directoryPath, { recursive: true });
+    //         }
+    //     }
+    
+    //     await sharp(req.file.path)
+    //         .toFile(`./static/facadesItems/${req.file.originalname}`)
+    
+    //     const url = `https://frolfasd.ru/static/facadesItems/${req.file.originalname}`
+    //     console.log('изменение сработало!')
+    //     fs.unlink(req.file.path, () => { // Для удаления закодированных файлов после использования
+    //         console.log(req.file.path)
+    //     })
+    //     VentilatedFacadesItemsDAO.insertNew(url, ventilatedFacadeId)
+    //         .then((data) => {
+    //             res.json(data)
+    //         })
+    //         .catch((error) => {
+    //             if (error.status === 500) {
+    //                 res.status(500).send({ status: 'Problem', message: 'Problem with database' })
+    //             } else {
+    //                 res.status(400).send({ status: 'Bad Request', message: error.message })
+    //             }
+    //         });
+    // }
 
     async getVentilatedFacadeItemsFromOneVentilatedFacade(req, res) {
         const ventilatedFacadeId = req.params.id
